+++ src/main/java/de/jeff_media/bettertridents/config/Permissions.java
@@ -1,0 +1,6 @@
+package de.jeff_media.bettertridents.config;
+
+public class Permissions {
+    public static final String SAVE_VOID = "bettertridents.savevoid";
+    public static final String RELOAD = "bettertridents.reload";
+}