+++ src/main/java/de/jeff_media/notridentvoid/config/Permissions.java
@@ -1,5 +1,0 @@
-package de.jeff_media.notridentvoid.config;
-
-public class Permissions {
-    public static final String NOTRIDENTVOID_USE = "notridentvoid.use";
-}