+++ CHANGELOG.md
@@ -1,2 +1,12 @@
+## 2.0.0
+- Renamed plugin to "BetterTridents"
+- Added new mechanics and config options:
+  - Bedrock impaling: Impaling enchantment behaves like in Bedrock (it damages ALL mobs touching water, not only ocean mobs). Default: Enabled
+  - Bedrock drop chance: Raises the drop chance for tridents to the bedrock value (25% for every drowned yielding a trident + 4% per looting level) Default: Enabled
+  - Return to offhand: Tridents thrown from the offhand will also return to the offhand when picking them up. Default: Enabled
+  - Disable loyalty portals: Prohibit tridents with Loyalty to travel through portals to avoid losing them. Default: Enabled
+- Added automatic config updater
+- Added reload command (/bettertridents, permission bettertridents.reload)
+
 ## 1.1.0
 - Added UpdateChecker and bStats